Disable interrupts to use 1200 LEDs
Once the number of LEDs gets large, handling interrupts
while bitbanging the colors will cause weird artifacts.
This is because the line will stay low, leading the LEDs
to think that they should apply the settings seen so far.

Luckily, we can tell FastLED not to handle interrupts while
bitbanging. This is likely to be temporary as I intend to
switch OctoWS which uses the DMA.