Switch to OctoWS library
Use OctoWS to write to the LED using DMA, but use FastLEDs color
utilities. Turns out that we can just pass in the CRGB array as the
drawing memory so we don't even need FastLEDs OctoWS controller!

Wow... just wow. The speed gains are impressive.

The CPU load for the 1200 LEDs with FastLED was around 90%.
Now, thanks to OctoWS, it's 3.4%... insane!