Add rudimentary diagnostics mode
On the USB Serial port, send 'd' to enter diagnostics mode.
It'll play a short sequence to indicate that it's ready. After that
type 'L' and a number followed by enter to turn the corresponding LED on,
or type 'T' and a number followed by enter to turn the corresponding Triangle
on. Type 'x' to exit diagnostics mode.

This diagnostic mode has revealed that there's an offset of 1 LED on the
WS2815 compared to the short WS2813 strip I have. This could be due to several
reasons, most probably signal quality.