Add rudimentary diagnostics mode
On the USB Serial port, send 'd' to enter diagnostics mode.
It'll play a short sequence to indicate that it's ready. After that
type 'L' and a number followed by enter to turn the corresponding LED on,
or type 'T' and a number followed by enter to turn the corresponding Triangle
on. Type 'x' to exit diagnostics mode.