Allow LEDs to be indexed on a per strip basis
S3;22 would turn on LED 22 on strip 3.

If you don't enter a semicolon or colon, it would just turn on the given
LED on the last strip used, e.g.

S40 would turn on LED 40 on strip 3 if it followed the above command.
If there was no preceding command, the strip defaults to strip 0, so
S40 would turn on LED 40 on strip 0