Provide iterator functionality to allow first LED to be on corner that isn't the first corner.
In other words, regardless on which corner the first LED of the strip is located
the first LED returned by begin() is always the one starting on the first corner
on the first edge.

At this point the direction of increasing LED indices must match the direction
of the corners, e.g. if the corners are clock-wise, the LED strip must be
installed clock-wise too.