Allow strips to run in opposite direction of triangle
Add support for strips to run in the opposite direction of the
triangle, e.g. if the corners of the triangle are clockwise
and the strip runs counter-clockwise, the iterator still runs
clockwise. This means that patterns will have the same orientation
regardless of the direction the strip runs around the triangle