First take on diffusion
This brings the diffusion cloth that we span over the triangles into play.
I'm on the fence whether this is too specific to the dome project and
I should bother trying to keep backwards compatibility with the particle
approach. For now both still are supported.

Needs tidying up, and a whole lotta documenting. Ah, and the rendering
sucks with large window sizes, and full-screen is a shocker.