Support overlapping shapes
One would expect that overlapping shapes behave reasonably, i.e. if a key
for a key is pressed that creates a shape overlapping another, that shape is
fully visible while the other one is only partially visible (i.e. covered),
and once the key for the overlapping shape is released, the previously covered
shape is fully visible.

It's all fairly disgusting and in dire need of a cleanup, though.