Rework ShapesFromNotes
On second thought, having the blobs and pentagons
on separate channels is weird, so let's just use
one channel, with a second one for the individual triangle.

It's still a bit weird, and likely not the last change in this area.
I guess we'll have to work with it for a bit and see how we go.

Also, since we no longer have separate channels for the different shapes,
request the Komplete Kontrol colors explicitly