native: Add virtual output
We now have a 'GeoLEDic In' and a 'GeoLEDic Out' virtual port to connect
e.g. Ableton's output and input to. Make sure that data from any input
doesn't get echoed to the virtual output to prevent an endless loop.

This means we only send program changes and controller changes originating
within GeoLEDic to the virtual output.

It's also handy to send a snapshot on demand, and not only on resuming output.