Add better version of Edgy: BetterEdgy
This improved version uses center values for hue and saturation
plus the amount of randomised deviation. I think this is easier
to use than the max and min approach, plus it handles the fact
that the hue is circular much better.

Also, optionally use a separate set of color configs for the
keyboard-triggered shapes. When I did the 'Drawn out like an ache'
animation I had to change the parameters back and forth every time
I hit a key, which was cumbersome.

To avoid having to redo the above animation I decided to add BetterEdgy
rather than changing the existing program.